requirements.txt
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.2.0
streamlit==1.12.0
tensorflow==2.8.0
